refactor(Storybook): update configuration and styles after upgrade to v9
