+++ packages/beeq/.storybook/main.ts
@@ -1,10 +1,13 @@
-import { join } from 'path';
+import { createRequire } from 'node:module';
+import { dirname, join } from 'path';
 
 import type { StorybookConfig } from '@storybook/web-components-vite';
 
+const require = createRequire(import.meta.url);
+
 export default {
   framework: {
-    name: '@storybook/web-components-vite',
+    name: getAbsolutePath('@storybook/web-components-vite'),
     options: {
       builder: {
         viteConfigPath: join(__dirname, '../vite.config.mts').replace(/\\/g, '/'),
@@ -13,10 +16,10 @@ export default {
   },
   stories: ['../src/_storybook/**/*.mdx', '../src/**/*.stories.@(ts|tsx)'],
   addons: [
-    '@storybook/addon-essentials',
-    '@storybook/addon-a11y',
-    '@beeq/storybook-addon-html',
-    '@chromatic-com/storybook',
+    getAbsolutePath('@storybook/addon-docs'),
+    getAbsolutePath('@storybook/addon-a11y'),
+    // getAbsolutePath('@beeq/storybook-addon-html'),
+    getAbsolutePath('@chromatic-com/storybook'),
   ],
   staticDirs: [
     { from: '../../../dist/beeq/www/assets', to: '/assets' },
@@ -46,3 +49,7 @@ export default {
     </script>
   `,
 } satisfies StorybookConfig;
+
+function getAbsolutePath(value: string): string {
+  return dirname(require.resolve(join(value, 'package.json')));
+}
